// types/auth/Note.ts
import Pagination from './Pagination';

export type Note = {
  id: string;
  title: string;
  body: string;
  verseKey?: string; // will be undefined when the note is not attached to any verse
  ranges?: string[]; // will be undefined when the note is not attached to any verse
  createdAt: Date;
  updatedAt: Date;
};

export type GetAllNotesResponse = {
  data: Note[];
  pagination: Pagination;
};